refactor: standardize file naming convention for type suffixes
