+++ packages/docs-ui/src/plugin.ts
@@ -120,7 +120,7 @@ import { DocMenuStyleService } from './services/doc-menu-style.service';
 import { DocPageLayoutService } from './services/doc-page-layout.service';
 import { DocParagraphMenuService } from './services/doc-paragraph-menu.service';
 import { DocCanvasPopManagerService } from './services/doc-popup-manager.service';
-import { DocPrintInterceptorService } from './services/doc-print-interceptor-service';
+import { DocPrintInterceptorService } from './services/doc-print-interceptor.service';
 import { DocStateChangeManagerService } from './services/doc-state-change-manager.service';
 import { DocsRenderService } from './services/docs-render.service';
 import { EditorService, IEditorService } from './services/editor/editor-manager.service';
