+++ packages/sheets-numfmt-ui/src/plugin.ts
@@ -23,7 +23,7 @@ import { UniverSheetsUIPlugin } from '@univerjs/sheets-ui';
 import pkg from '../package.json';
 import { defaultPluginConfig } from './config/config';
 import { NumfmtAlertRenderController } from './controllers/numfmt-alert-render.controller';
-import { NumfmtRepeatLastActionController } from './controllers/numfmt-repeat-last-action-controller';
+import { NumfmtRepeatLastActionController } from './controllers/numfmt-repeat-last-action.controller';
 import { SheetNumfmtUIController } from './controllers/numfmt.controller';
 import { NumfmtEditorController } from './controllers/numfmt.editor.controller';
 import { UserHabitController } from './controllers/user-habit.controller';
