+++ packages/sheets-table-ui/src/plugin.ts
@@ -24,14 +24,14 @@ import { OpenTableFilterPanelOperation } from './commands/operations/open-table-
 import { OpenTableSelectorOperation } from './commands/operations/open-table-selector.operation';
 import { defaultPluginConfig, SHEETS_TABLE_UI_PLUGIN_CONFIG_KEY } from './config/config';
 import { PLUGIN_NAME } from './const';
-import { SheetTableControlsRenderController } from './controllers/sheet-table-controls-render.controller';
 import { SheetsTableComponentController } from './controllers/sheet-table-component.controller';
+import { SheetTableControlsRenderController } from './controllers/sheet-table-controls-render.controller';
 import { SheetsTableFilterButtonRenderController } from './controllers/sheet-table-filter-button-render.controller';
 import { SheetsTableRenderController } from './controllers/sheet-table-render.controller';
 import { SheetTableSelectionController } from './controllers/sheet-table-selection.controller';
 import { SheetTableThemeUIController } from './controllers/sheet-table-theme-ui.controller';
 import { SheetTableMenuController } from './menu/sheet-table-menu.controller';
-import { SheetsTableUiService } from './services/sheets-table-ui-service';
+import { SheetsTableUiService } from './services/sheets-table-ui.service';
 
 @DependentOn(UniverSheetsTablePlugin)
 export class UniverSheetsTableUIPlugin extends Plugin {
