+++ packages/telemetry/src/index.spec.ts
@@ -16,7 +16,7 @@
 
 import { describe, expect, it } from 'vitest';
 import * as indexModule from './index';
-import { ITelemetryService } from './services/telemetry-service';
+import { ITelemetryService } from './services/telemetry.service';
 
 describe('telemetry exports', () => {
     it('should re-export ITelemetryService identifier from index', () => {
